Set fd limits and begin io_uring
Set the file descriptor limits to their absolute maximum we can go as
there will be a lot of open files/folders while these operations are
performed on them. Got the basics of io_uring setup and ready, next
step should be writing the directory traversal part and the thread pool.