Restructure how events are processed
Because directories are complicated, this now processes everything in
separate threads which makes it truly asynchronous. I need to make a
recursive asynchronous mkdir implementation next before I can build the
rest of the code to copy symlinks and files. So far the application
should descend all subdirectories (in theory).