First attempt at recursive asynchronous mkdir
This crashes due to some memory problems but it should, in theory,
create directories (and sub directories) for a given directory path.
This should be the async equivalent of doing `mkdir -p <some path>`.

I need to debug why it crashes and sometimes doesn't produce the proper
paths as expected (missing the copy-to path prefix, for example).